up：修改返回值类型.

+++ easy_net/base/thread.h
@@ -12,9 +12,9 @@
 namespace EasyNet {
 inline size_t GetCurrentThreadId() {
 #ifdef _WIN32
-    return static_cast<long long>(::GetCurrentThreadId());
+    return static_cast<size_t>(::GetCurrentThreadId());
 #elif defined(__linux__)
-    return static_cast<long long>(syscall(SYS_gettid));
+    return static_cast<size_t>(syscall(SYS_gettid));
 #else // Default to standard C++11 (other Unix)
     return static_cast<size_t>(std::hash<std::thread::id>()(std::this_thread::get_id()));
 #endif
